Revert "YDM-28 // changed UserId to Long - before changing AddMovieToListSample"
This reverts commit 9c4c6808659a0d36fd4160712210acd06c6bbf31.

+++ src/main/java/pl/szczesniak/dominik/whattowatch/users/domain/model/UserId.java
@@ -15,10 +15,10 @@ import static com.google.common.base.Preconditions.checkArgument;
 @EqualsAndHashCode
 public class UserId {
 
-	private Long value;
+	private Integer value;
 
-	public UserId(final Long value) {
-		checkArgument(value > 0L, "UserId value must be higher than 0");
+	public UserId(final int value) {
+		checkArgument(value > 0, "UserId value must be higher than 0");
 		this.value = value;
 	}
