+++ src/test/java/pl/szczesniak/dominik/whattowatch/movies/FindMoviesToWatchRestInvoker.java
@@ -23,7 +23,7 @@ public class FindMoviesToWatchRestInvoker {
 
 	private final TestRestTemplate restTemplate;
 
-	public ResponseEntity<List<MovieDto>> findMoviesToWatch(final Long userId) {
+	public ResponseEntity<List<MovieDto>> findMoviesToWatch(final Integer userId) {
 		final HttpHeaders headers = new HttpHeaders();
 		headers.set("userId", String.valueOf(userId));
 		return restTemplate.exchange(
@@ -43,7 +43,7 @@ public class FindMoviesToWatchRestInvoker {
 
 		private final String title;
 		private final Integer movieId;
-		private final Long userId;
+		private final Integer userId;
 
 	}
